feat(cli): allow overwrite in some cases by default
Allow overwrite of the existing .rules file when it's empty, does not contain Svarog or includes a
compatible version of Svarog. Only generated portion will be overwritten an all cases.

BREAKING CHANGE: This commit changes the behavior of the --force flag: from now on it will allow
overwrite of the Svarog code only, not the entire file.